cloud-functions/cfpb-fetcher/requirements.txt
functions-framework==3.4.0
google-cloud-storage==2.10.0
requests==2.31.0
flask==2.3.3
